api/requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
redis
rq
beautifulsoup4
lxml
python-dotenv
sqlalchemy
pandas
numpy

